+++ changelog.md
@@ -61,7 +61,7 @@
 - Added `wrapnils` module for chains of field-access and indexing where the LHS can be nil.
   This simplifies code by reducing need for if-else branches around intermediate maybe nil values.
   E.g. `echo ?.n.typ.kind`
-- Added `minIndex` and `maxIndex` to the `sequtils` module
+- Added `minIndex`, `maxIndex` and `unzip` to the `sequtils` module.
 - Added `os.isRelativeTo` to tell whether a path is relative to another
 - Added `resetOutputFormatters` to `unittest`